Modified the storage layer. Add some ui stuff.
git-svn-id: https://rabbit-eclipse.googlecode.com/svn/trunk@19 03b96cd2-0a55-11df-9859-c12d74d802c5

+++ rabbit.tracking.ui.pages/src/rabbit/tracking/ui/pages/CommandPageLabelProvider.java
@@ -33,7 +33,7 @@ public class CommandPageLabelProvider extends LabelProvider implements ITableLab
 			case 1:
 				return cmd.getDescription();
 			case 2:
-				return Integer.toString(page.getUsage(cmd));
+				return Long.toString(page.getUsage(cmd));
 			default:
 				return null;
 			}
